refactor(storage): internal queue storage
