+++ collator/src/internal_queue/state/shard_iterator.rs
@@ -1,7 +1,7 @@
 use anyhow::Result;
 use everscale_types::models::ShardIdent;
 use tycho_block_util::queue::QueueKey;
-use tycho_storage::InternalQueueMessagesIter;
+use tycho_storage::iterator::InternalQueueMessagesIter;
 
 use crate::types::ShortAddr;
