+++ collator/src/internal_queue/state/state_iterator.rs
@@ -7,7 +7,7 @@ use ahash::HashMapExt;
 use anyhow::{bail, Context, Result};
 use everscale_types::models::ShardIdent;
 use tycho_block_util::queue::QueueKey;
-use tycho_storage::InternalQueueMessagesIter;
+use tycho_storage::iterator::InternalQueueMessagesIter;
 use tycho_util::FastHashMap;
 
 use crate::internal_queue::state::shard_iterator::{IterResult, ShardIterator};
